Wrangle the df and keep only the interested variables (unquoted) and output should have same type as of df.

Wrangle the df and keep only the interested variables (unquoted) and output should have same type as of df.

wrangle_data(df, ...)

Arguments

df

A data frame that needs to be wrangle

...

unquoted variables names to select data

Value

A data frame that contains only those variables specified

Examples

wrangle_data(mtcars, mpg)
#>                      mpg
#> Mazda RX4           21.0
#> Mazda RX4 Wag       21.0
#> Datsun 710          22.8
#> Hornet 4 Drive      21.4
#> Hornet Sportabout   18.7
#> Valiant             18.1
#> Duster 360          14.3
#> Merc 240D           24.4
#> Merc 230            22.8
#> Merc 280            19.2
#> Merc 280C           17.8
#> Merc 450SE          16.4
#> Merc 450SL          17.3
#> Merc 450SLC         15.2
#> Cadillac Fleetwood  10.4
#> Lincoln Continental 10.4
#> Chrysler Imperial   14.7
#> Fiat 128            32.4
#> Honda Civic         30.4
#> Toyota Corolla      33.9
#> Toyota Corona       21.5
#> Dodge Challenger    15.5
#> AMC Javelin         15.2
#> Camaro Z28          13.3
#> Pontiac Firebird    19.2
#> Fiat X1-9           27.3
#> Porsche 914-2       26.0
#> Lotus Europa        30.4
#> Ford Pantera L      15.8
#> Ferrari Dino        19.7
#> Maserati Bora       15.0
#> Volvo 142E          21.4
wrangle_data(mtcars, mpg, disp)
#>                      mpg  disp
#> Mazda RX4           21.0 160.0
#> Mazda RX4 Wag       21.0 160.0
#> Datsun 710          22.8 108.0
#> Hornet 4 Drive      21.4 258.0
#> Hornet Sportabout   18.7 360.0
#> Valiant             18.1 225.0
#> Duster 360          14.3 360.0
#> Merc 240D           24.4 146.7
#> Merc 230            22.8 140.8
#> Merc 280            19.2 167.6
#> Merc 280C           17.8 167.6
#> Merc 450SE          16.4 275.8
#> Merc 450SL          17.3 275.8
#> Merc 450SLC         15.2 275.8
#> Cadillac Fleetwood  10.4 472.0
#> Lincoln Continental 10.4 460.0
#> Chrysler Imperial   14.7 440.0
#> Fiat 128            32.4  78.7
#> Honda Civic         30.4  75.7
#> Toyota Corolla      33.9  71.1
#> Toyota Corona       21.5 120.1
#> Dodge Challenger    15.5 318.0
#> AMC Javelin         15.2 304.0
#> Camaro Z28          13.3 350.0
#> Pontiac Firebird    19.2 400.0
#> Fiat X1-9           27.3  79.0
#> Porsche 914-2       26.0 120.3
#> Lotus Europa        30.4  95.1
#> Ford Pantera L      15.8 351.0
#> Ferrari Dino        19.7 145.0
#> Maserati Bora       15.0 301.0
#> Volvo 142E          21.4 121.0